bugfix 2

+++ wp-content/plugins/medma-site-manager/includes/ordermap.class.php
@@ -6,18 +6,22 @@ class OrderMap {
         return $wpdb->base_prefix . 'order_map';
     }
 
-    public static function addNewRelation($user_id, $blog_id, $counter_id) {
+    public static function addNewRelation($user_id, $blog_id, $order_id) {
         global $wpdb;
-        $order_maping = $wpdb->insert(self::tableName(), array(
-            'user_id' => $user_id,
-            'site_id' => $blog_id,
-            'counter_id' => $counter_id
-        ), array(
-                '%d',
-                '%d',
-                '%d'
-            )
-        );
+
+        $counterTable = $wpdb->base_prefix . 'package_counter';
+        $counter_id = $wpdb->get_var('SELECT `id` FORM `'.$counterTable.'` WHERE `order_id` = '.(int)$order_id);
+        if ($counter_id && $wpdb->query('UPDATE `'.$counterTable.'` SET `site_consumed` = `site_consumed` + 1 WHERE `id` = '.(int)$counter_id )) {   $order_maping = $wpdb->insert(self::tableName(), array(
+                'user_id' => $user_id,
+                'site_id' => $blog_id,
+                'counter_id' => $counter_id
+            ), array(
+                    '%d',
+                    '%d',
+                    '%d'
+                )
+            );
+        }
     }
 
     public static function getBlogOwner($blog_id) {
@@ -62,6 +66,9 @@ class OrderMap {
     public static function dropRelation($blog_id) {
         /** @var wpdb */
         global $wpdb;
-        return $wpdb->delete($wpdb->base_prefix. self::tableName(), 'blog_id = '.(int)$blog_id);
+        $blogInfo = self::getBlogInfo($blog_id);
+        if ($blogInfo && $wpdb->query('UPDATE `'.$wpdb->base_prefix . 'package_counter` SET `site_consumed` = `site_consumed` - 1 WHERE `id` = '.$blogInfo->counter_id)) {
+            $wpdb->delete($wpdb->base_prefix . self::tableName(), 'blog_id = ' . (int)$blog_id);
+        }
     }
 }
